lex: consolidate integer literal parsing into lexer
Consolidates the parseInt function in integeralLiteral.zig into the
lexer, which now keeps track of the most recent integer literal parsed
(which is consumed by the parser). This avoids two codepaths which is
buggy and slower than necessary.

This fixes #54.

The only potential downside here is that its more difficult to emit
specific errors for different types of literal parsing errors, since we
don't have access to the parser's error infrastructure in lexer. To be
investigated.